requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
appdirs==1.4.4
    # via black
argon2-cffi==20.1.0
    # via -r requirements.in
asgiref==3.3.4
    # via django
black==20.8b1
    # via -r requirements.in
bleach==3.3.0
    # via -r requirements.in
cffi==1.14.4
    # via
    #   argon2-cffi
    #   cryptography
click==7.1.2
    # via
    #   black
    #   pip-tools
cryptography==3.4.3
    # via pyjwt
django-cors-headers==3.7.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-redis==5.0.0
    # via -r requirements.in
django==3.2.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-redis
    #   djangorestframework
    #   drf-jwt
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-jwt
drf-jwt==1.19.1
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==6.0.0
    # via factory-boy
flake8-blind-except==0.2.0
    # via -r requirements.in
flake8-builtins==1.5.3
    # via -r requirements.in
flake8-commas==2.0.0
    # via -r requirements.in
flake8-debugger==4.0.0
    # via -r requirements.in
flake8-mutable==1.2.0
    # via -r requirements.in
flake8-polyfill==1.0.2
    # via pep8-naming
flake8-tuple==0.4.1
    # via -r requirements.in
flake8==3.9.2
    # via
    #   -r requirements.in
    #   flake8-builtins
    #   flake8-commas
    #   flake8-debugger
    #   flake8-mutable
    #   flake8-polyfill
    #   flake8-tuple
marshmallow==3.12.1
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via bleach
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
pep8-naming==0.11.1
    # via -r requirements.in
pip-tools==6.1.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-debugger
pycparser==2.20
    # via cffi
pyflakes==2.3.1
    # via flake8
pyjwt[crypto]==1.7.1
    # via drf-jwt
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via faker
pytz==2021.1
    # via
    #   -r requirements.in
    #   django
redis==3.5.3
    # via
    #   -r requirements.in
    #   django-redis
regex==2020.11.13
    # via black
six==1.15.0
    # via
    #   argon2-cffi
    #   bleach
    #   flake8-debugger
    #   flake8-tuple
    #   python-dateutil
sqlparse==0.4.1
    # via django
tblib==1.7.0
    # via -r requirements.in
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   pep517
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via black
webencodings==0.5.1
    # via bleach

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
